Changed application to use Quarkus instead of Wildfly
