--- api-client/build.gradle
@@ -1,42 +1,0 @@
-buildscript {
-    repositories {
-        jcenter()
-    }
-}
-
-plugins {
-    id "org.openapi.generator" version "4.2.3"
-}
-
-apply plugin: 'kotlin'
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    compile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    compile "com.squareup.moshi:moshi-kotlin:1.8.0"
-    compile "com.squareup.moshi:moshi-adapters:1.8.0"
-    compile "com.squareup.okhttp3:okhttp:4.2.2"
-}
-
-openApiGenerate {
-    generatorName = "kotlin"
-    inputSpec = "$rootDir/megasense-data-portal-api-spec/swagger.yaml".toString()
-    outputDir = "$buildDir/generated".toString()
-    packageName = "fi.metatavu.megasense.dataportal.api.client"
-    configOptions = [
-            library: "jvm-okhttp4",
-            dateLibrary: "string"
-    ]
-}
-
-task copyGeneratedCode(type: Copy) {
-    from "$buildDir/generated/src/main"
-    into "src/main"
-    dependsOn tasks.openApiGenerate
-}
-
-compileKotlin.dependsOn tasks.copyGeneratedCode
