--- docker-compose.yml
@@ -1,64 +1,0 @@
-version: "3"
-services:
-  keycloak-mysql:
-    image: mysql:5.7
-    environment:
-      MYSQL_ROOT_PASSWORD: root
-      MYSQL_DATABASE: keycloak
-      MYSQL_USER: keycloak
-      MYSQL_PASSWORD: password
-  megasense-data-portal-api-mysql:
-    image: mysql:5.7
-    command: --lower_case_table_names=1
-    environment:
-      MYSQL_ROOT_PASSWORD: root
-      MYSQL_DATABASE: megasense-data-portal-api
-      MYSQL_USER: megasense-data-portal-api
-      MYSQL_PASSWORD: password
-  test-keycloak:
-    image: jboss/keycloak:8.0.1
-    environment:
-      DB_VENDOR: MYSQL
-      DB_ADDR: keycloak-mysql
-      DB_DATABASE: keycloak
-      DB_USER: keycloak
-      DB_PASSWORD: password
-      JDBC_PARAMS: "verifyServerCertificate=false&useSSL=false"
-      KEYCLOAK_USER: admin
-      KEYCLOAK_PASSWORD: admin
-      KEYCLOAK_IMPORT: /opt/docker/kc.json
-    ports:
-      - 8080:8080
-    depends_on:
-      - keycloak-mysql
-    volumes:
-      - ./test-volumes/keycloak:/opt/docker
-  megasense-data-portal-api:
-    build: .
-    ports:
-      - 1234:8080
-    depends_on:
-      - megasense-data-portal-api-mysql
-      - test-keycloak
-    environment:
-      API_HOST: localhost
-      API_CLUSTER_PASSWORD: dfd9f1b6-3b04-4777-9c21-fad19ef5b8d3
-      API_DB_HOST: megasense-data-portal-api-mysql
-      API_DB_NAME: megasense-data-portal-api
-      API_DB_USERNAME: megasense-data-portal-api
-      API_DB_PASSWORD: password
-      KEYCLOAK_SECRET: f20e3183-34fa-4496-8282-7dcf177ca09b
-      KEYCLOAK_REALM: megasense-data-portal
-      KEYCLOAK_RESOURCE: api
-      KEYCLOAK_URL: http://test-keycloak:8080/auth
-      KEYCLOAK_USER: admin
-      KEYCLOAK_PASSWORD: admin
-      KEYCLOAK_ADMIN_CLIENT_ID: admin-cli
-      KUBERNETES_NAMESPACE: test
-      FILE_STORAGE_PROVIDER: LOCAL
-      LOCAL_FILE_STORAGE_PATH: /opt/uploads
-      JAVA_OPTS: '-javaagent:/opt/jacoco-jacocoagent.jar=output=file,append=false,dumponexit=true,destfile=/opt/jacoco/it.exec'
-    volumes:
-      - ./test-volumes/uploads:/opt/uploads
-      - $JACOCO_AGENT:/opt/jacoco-jacocoagent.jar
-      - $JACOCO:/opt/jacoco
