--- docker-resources/host.cli
@@ -1,2 +1,0 @@
-embed-server --server-config=standalone-full-ha.xml
-/subsystem=undertow/server=default-server/host=megasense-data-portal-api:add(alias=["${env.API_HOST}"],default-web-module="megasense-data-portal-api.war")
