--- docker-resources/jboss-cli.properties
@@ -1,1 +1,0 @@
-server.config=standalone-full-ha.xml
