--- docker-resources/jdbc.cli
@@ -1,3 +1,0 @@
-embed-server --server-config=standalone-full-ha.xml
-/subsystem=datasources/jdbc-driver=mysql:add(driver-name=mysql, driver-module-name=com.mysql.jdbc, xa-datasource-class=com.mysql.jdbc.jdbc2.optional.MysqlXADataSource)
-data-source add --name=megasense-data-portal-api --jndi-name=java:jboss/datasources/megasense-data-portal-api --enabled=true --jta=true --use-ccm=true --connection-url=jdbc:mysql://${env.API_DB_HOST}:${env.API_DB_PORT:3306}/${env.API_DB_NAME}${env.API_DB_OPTS:} --driver-name=mysql --user-name=${env.API_DB_USERNAME} --password=${env.API_DB_PASSWORD} --check-valid-connection-sql="SELECT 1" --validate-on-match=false --share-prepared-statements=false --max-pool-size=1000
