--- docker-resources/keycloak.cli
@@ -1,8 +1,0 @@
-embed-server --server-config=standalone-full-ha.xml
-
-/subsystem=security/security-domain=keycloak/:add
-/subsystem=security/security-domain=keycloak/authentication=classic/:add(login-modules=[{ "code" => "org.keycloak.adapters.jboss.KeycloakLoginModule","flag" => "required"}])
-/subsystem=logging/logger=org.keycloak.adapters.BearerTokenRequestAuthenticator:add(level=FATAL)
-/subsystem=logging/logger=org.jboss.as.ejb3.invocation:add(level=FATAL)
-/subsystem=keycloak/secure-deployment=megasense-data-portal-api.war:add(realm=${env.KEYCLOAK_REALM},resource=${env.KEYCLOAK_RESOURCE},bearer-only=true,auth-server-url=${env.KEYCLOAK_URL},ssl-required=EXTERNAL,enable-cors=true)
-/subsystem=keycloak/secure-deployment=megasense-data-portal-api.war/credential=secret:add(value=${env.KEYCLOAK_SECRET})
