+++ gradle.properties
@@ -1,1 +1,8 @@
-systemProp.jdk.tls.client.protocols="TLSv1,TLSv1.1,TLSv1.2"
+#Gradle properties
+#Wed Feb 10 09:13:25 EET 2021
+quarkusPluginVersion=1.10.5.Final
+quarkusPlatformArtifactId=quarkus-universe-bom
+quarkusPluginId=io.quarkus
+quarkusPlatformGroupId=io.quarkus
+quarkusPlatformVersion=1.10.5.Final
+org.gradle.logging.level=INFO
