--- src/main/kotlin/fi/metatavu/megasense/dataportal/api/Application.kt
@@ -1,11 +1,0 @@
-package fi.metatavu.megasense.dataportal.api
-
-import javax.ws.rs.ApplicationPath
-import javax.ws.rs.core.Application
-
-/**
- * JAX-RS Activator class
- */
-@ApplicationPath("/v1")
-open class JaxRsActivator : Application() {
-}
