+++ src/main/kotlin/fi/metatavu/megasense/dataportal/resteasyjackson/MyObjectMapperCustomizer.kt
@@ -1,0 +1,15 @@
+package fi.metatavu.megasense.dataportal.resteasyjackson
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.SerializationFeature
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import io.quarkus.jackson.ObjectMapperCustomizer
+import javax.inject.Singleton
+
+@Singleton
+class MyObjectMapperCustomizer : ObjectMapperCustomizer {
+    override fun customize(objectMapper: ObjectMapper) {
+        objectMapper.registerModule(JavaTimeModule())
+        objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+    }
+}
