+++ src/main/resources/application.properties
@@ -1,0 +1,11 @@
+# http configuration
+
+# datasource configuration
+quarkus.datasource.db-kind = mysql
+quarkus.datasource.username = ${JDBC_USER}
+quarkus.datasource.password = ${JDBC_PASSWORD}
+quarkus.datasource.jdbc.url = ${JDBC_URL}
+
+# Liquibase minimal config properties
+quarkus.liquibase.migrate-at-start=true
+quarkus.http.cors=true
