+++ src/test/kotlin/fi/metatavu/megasense/dataportal/api/test/functional/builder/impl/FavouritesTestBuilderResource.kt
@@ -28,7 +28,7 @@ class FavouritesTestBuilderResource(testBuilder: AbstractTestBuilder<ApiClient?>
      *
      * @return all favourites created by the user
      */
-    fun listAll (): Array<FavouriteLocation> {
+    fun listAll (): List<FavouriteLocation> {
         return api.listUserFavouriteLocations()
     }
