+++ src/test/resources/kc.json
